% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{somar}
\alias{somar}
\title{Function to sum}
\usage{
somar(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It's another number.}
}
\description{
This function is going to sum.
}
\examples{
sum (2, 3)

}
