% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{subtrair}
\alias{subtrair}
\title{Function to subtract}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It's another number.}
}
\description{
This function is going to subtract.
}
\examples{
subtrair (2, 3)

}
